Classes created and bugs vector populated
Derived Hopper and Crawler classes made with appropriate values.

tokenizeInputStream()
Function accepts a line from the txt, the type of bug and a reference to bugs vector. Using a delimiter it captures the value found and adds to an attributes vector. Depending on the type passed a new Hopper or Crawler class is created with the attributes collected from the line