Redesign of move and eat functionality
Move now has more of a randomness to it
Simulation works as follows:
Bugs are loaded into bugs_vector from txt
When the simulation is ran the board is tapped repeatedly. This calls updateBoard which sets all posiitions to nullptrs then sets positions again with posiitons from bug_vector

Inside tapBugBoard bugs are looped to call their move functions. If two bugs land on the same position they will fight.

Back to the top of the loop where updateBoard is called again.  All points on the board are set to nullptrs. If a bug has alive value = true, they will be added to the board

This is repeated until 1 bug remains